remove remaining integer, replace with bigint

+++ pedsnetdcc/bmi.py
@@ -347,7 +347,7 @@ def _add_measurement_ids(conn_str, site, neg_ids, search_path, model_version, id
     lock_last_id_msg = "locking {table_name} last ID tracking table for update"
 
     update_last_id_sql = """UPDATE {last_id_table_name} AS new
-        SET last_id = new.last_id + '{new_id_count}'::integer
+        SET last_id = new.last_id + '{new_id_count}'::bigint
         FROM {last_id_table_name} AS old RETURNING old.last_id, new.last_id"""
     update_last_id_msg = "updating {table_name} last ID tracking table to reserve new IDs"  # noqa
     create_seq_measurement_sql = "create sequence if not exists {0}.{1}_bmi_measurement_id_seq"
